+++ src/main/java/com/sprint/part2/sb1hrbankteam03/dto/backup/BackupDto.java
@@ -1,4 +1,4 @@
-package com.sprint.part2.sb1hrbankteam03.dto;
+package com.sprint.part2.sb1hrbankteam03.dto.backup;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
