+++ src/main/java/com/sprint/part2/sb1hrbankteam03/dto/backup/RequestBackupDto.java
@@ -1,4 +1,4 @@
-package com.sprint.part2.sb1hrbankteam03.dto;
+package com.sprint.part2.sb1hrbankteam03.dto.backup;
 
 import lombok.Getter;
